extract test code to separate file
